make component abstract
